# contracts/openapi/alignment-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Traffic Analysis Platform Alignment API",
    "version": "1.2.0",
    "description": "Additive contract for the four remediation vertical slices. Existing APIs remain available during migration."
  },
  "servers": [{"url": "/api"}],
  "paths": {
...
    "/v1/dashboard/tasks": {
      "post": {
        "operationId": "createDashboardTask",
        "x-feature-id": "F-DASHBOARD-002",
        "x-required-scope": "dashboard:write",
        "parameters": [{"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DashboardTaskCreate"}}}},
        "responses": {
          "202": {"description": "Task, immutable history, minimum audit and outbox durably accepted; this is not final completion", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
          "403": {"description": "dashboard:write scope required"},
          "409": {"description": "Idempotency key reused with a different command"},
          "503": {"description": "Feature, PostgreSQL or versioned schema unavailable"}
        }
      }
    },
...
    "/v1/dashboard/tasks/{task_id}": {
      "get": {
        "operationId": "getDashboardTask", "x-feature-id": "F-DASHBOARD-002", "x-required-scope": "dashboard:write",
        "parameters": [{"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}],
        "responses": {
          "200": {"description": "Authoritative tenant-scoped accepted, running, completed, partial, failed or cancelled state", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
          "403": {"description": "dashboard:write scope required"},
          "404": {"description": "Task is absent from the authenticated tenant"}
        }
      }
    },
    "/v1/dashboard/tasks/{task_id}/compensations": {
      "post": {
        "operationId": "compensateDashboardTask",
        "x-feature-id": "F-DASHBOARD-002",
        "x-required-scope": "dashboard:write",
        "parameters": [
          {"name": "task_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}},
          {"$ref": "#/components/parameters/IdempotencyKey"}
        ],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DashboardTaskCompensationCreate"}}}},
        "responses": {
          "202": {"description": "Provider compensation durably accepted; final success still requires a confirmed provider receipt", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Envelope"}}}},
          "403": {"description": "dashboard:write scope required"},
          "404": {"description": "Task or confirmed provider receipt is absent from the authenticated tenant"},
          "409": {"description": "Idempotency, revision or task-state conflict"},
          "503": {"description": "Compensation pipeline, provider or versioned schema unavailable"}
        }
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 16}},
...
    },
...
    "schemas": {
...
      "DashboardTaskCreate": {
        "type": "object",
        "required": ["target", "priority", "snapshot_id", "reason"],
        "properties": {
          "target": {"type": "string", "minLength": 1, "maxLength": 500},
          "priority": {"enum": ["low", "medium", "high", "critical"]},
          "snapshot_id": {"type": "string", "minLength": 1, "maxLength": 500},
          "reason": {"type": "string", "minLength": 1, "maxLength": 2000},
          "context": {"type": "object", "additionalProperties": true}
        }
      },
      "DashboardTaskCompensationCreate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["action_id", "expected_revision", "reason"],
        "properties": {
          "action_id": {"const": "dashboard-task-compensate"},
          "expected_revision": {"type": "integer", "minimum": 1},
          "reason": {"type": "string", "minLength": 8, "maxLength": 2000}
        }
      },
      "Envelope": {
        "type": "object",
        "required": ["data", "meta", "error"],
        "properties": {
          "data": {},
          "meta": {"$ref": "#/components/schemas/Meta"},
          "error": {"oneOf": [{"type": "null"}, {"$ref": "#/components/schemas/Error"}]}
        }
      },
      "Meta": {
        "type": "object",
        "required": ["contract_version", "schema_version", "snapshot_id", "as_of", "generated_at", "trace_id", "result_code", "partial", "missing_sections", "source_watermarks"],
        "properties": {
          "contract_version": {"type": "integer"},
          "schema_version": {"type": "integer"},
          "snapshot_id": {"type": "string"},
          "as_of": {"type": "string", "format": "date-time"},
          "generated_at": {"type": "string", "format": "date-time"},
          "trace_id": {"type": "string"},
          "result_code": {"type": "string"},
          "operation_id": {"type": "string"},
          "tenant_id": {"type": "string"},
          "partial": {"type": "boolean"},
          "missing_sections": {"type": "array", "items": {"type": "string"}},
          "source_watermarks": {"type": "object", "additionalProperties": {"type": "string"}},
          "projection_status": {"type": "string"}
        }
      },
      "Error": {
        "type": "object",
        "required": ["code", "message", "trace_id", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "trace_id": {"type": "string"},
          "retryable": {"type": "boolean"},
          "field_errors": {"type": "array", "items": {"type": "object", "required": ["field", "code", "message"], "properties": {"field": {"type": "string"}, "code": {"type": "string"}, "message": {"type": "string"}}}},
          "operation_id": {"type": "string"},
          "current_revision": {"type": "integer"},
          "projection_status": {"type": "string"},
          "status_url": {"type": "string"},
          "expected_consistency_window": {"type": "string"}
        }
      },
...
    }
  }
}